backend/requirements.txt
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Supabase Client
supabase==2.3.4

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Geospatial Calculations
geopy==2.4.1

# Google Maps API
googlemaps==4.10.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Date/Time Handling
python-dateutil==2.8.2

# JSON Processing
jsonschema==4.20.0